requirements.txt
torch==1.7.1
torchvision==0.8.2
opencv-python==4.5.4.60
numpy==1.15.1
Pillow==8.4.0
tqdm
matplotlib